requirements.txt
google-generativeai
sacrebleu
unbabel-comet
transformers
sentence-transformers
torch
pypdf
python-docx
python-pptx